Move TranscodingJobDto and TranscodingThrottler to Controller
