+++ Jellyfin.Api/Helpers/FileStreamResponseHelpers.cs
@@ -4,7 +4,6 @@ using System.Net.Http;
 using System.Net.Mime;
 using System.Threading;
 using System.Threading.Tasks;
-using Jellyfin.Api.Models.PlaybackDtos;
 using Jellyfin.Api.Models.StreamingDtos;
 using MediaBrowser.Controller.MediaEncoding;
 using Microsoft.AspNetCore.Http;
